+++ cointrader/cli.py
@@ -1,11 +1,11 @@
 # -*- coding: utf-8 -*-
 import click
 import sys
-import datetime
 import logging
+import datetime
 from cointrader import db, STRATEGIES
 from cointrader.config import Config, get_path_to_config
-from cointrader.exchange import Poloniex
+from cointrader.exchange import Poloniex, BacktestMarket, Market
 from cointrader.exchanges.poloniex import ApiError
 from cointrader.bot import init_db, get_bot
 from cointrader.helpers import render_bot_statistic, render_bot_tradelog
@@ -44,7 +44,7 @@ def main(ctx, config):
 @click.command()
 @click.option("--order-by-volume", help="Order markets by their trading volume", is_flag=True)
 @click.option("--order-by-profit", help="Order markets by their current profit", is_flag=True)
-@click.option("--limit", help="Limit output to NUM markets", default=3)
+@click.option("--limit", help="Limit output to NUM markets", default=10)
 @pass_context
 def explore(ctx, order_by_volume, order_by_profit, limit):
     """List top markets. On default list markets which are profitable and has a high volume."""
@@ -54,16 +54,22 @@ def explore(ctx, order_by_volume, order_by_profit, limit):
         for market in markets:
             url = "https://poloniex.com/exchange#{}".format(market[0].lower())
             click.echo("{:<10} {:>6}% {:>10} {:>20}".format(market[0], market[1]["change"], market[1]["volume"], url))
+        if len(markets) == 0:
+            click.echo("Sorry. Can not find any market which is in the TOP{} for profit and trade volume. Try to increase the limit using the --limit parameter.".format(limit))
     elif order_by_volume:
         markets = ctx.exchange.get_top_volume_markets(markets, limit)
         for market in markets:
             url = "https://poloniex.com/exchange#{}".format(market[0].lower())
             click.echo("{:<10} {:>10} {:>6}% {:>20}".format(market[0], market[1]["volume"], market[1]["change"], url))
+        if len(markets) == 0:
+            click.echo("Sorry. Can not find any market which is in the TOP{} for trade volume. Try to increase the limit using the --limit parameter.".format(limit))
     elif order_by_profit:
         markets = ctx.exchange.get_top_profit_markets(markets, limit)
         for market in markets:
             url = "https://poloniex.com/exchange#{}".format(market[0].lower())
             click.echo("{:<10} {:>6}% {:>10} {:>20}".format(market[0], market[1]["change"], market[1]["volume"], url))
+        if len(markets) == 0:
+            click.echo("Sorry. Can not find any market which is in the TOP{} for profit. Try to increase the limit using the --limit parameter.".format(limit))
 
 
 @click.command()
@@ -94,26 +100,43 @@ def balance(ctx):
 @pass_context
 def start(ctx, market, resolution, start, end, automatic, backtest, papertrade, strategy, btc, coins):
     """Start a new bot on the given market and the given amount of BTC"""
-
+    # Check start and end date
     try:
-        market = ctx.exchange.get_market(market, backtest, papertrade)
-    except ApiError as ex:
-        click.echo(ex.message)
+        if start:
+            start = datetime.datetime.strptime(start, "%Y-%m-%d %H:%M:%S")
+        if end:
+            end = datetime.datetime.strptime(end, "%Y-%m-%d %H:%M:%S")
+    except ValueError:
+        click.echo("Date is not valid. Must be in format 'YYYY-mm-dd HH:MM:SS'")
         sys.exit(1)
-    except ValueError as ex:
-        click.echo(ex.message)
+
+    # Build the market on which the bot will operate
+    # First check if the given market is a valid market. If not exit
+    # here with a error message.
+    # If the market is valid create a real market instance of and
+    # instance for backtests depending on the user input.
+    if ctx.exchange.is_valid_market(market):
+        if backtest:
+            if start is None or end is None:
+                click.echo("Error! For backtests you must provide a timeframe by setting start and end!")
+                sys.exit(1)
+            market = BacktestMarket(ctx.exchange, market)
+        else:
+            market = Market(ctx.exchange, market)
+    else:
+        click.echo("Market {} is not available".format(market))
         sys.exit(1)
+
+    # Check if the given resolution is supported
+    if not ctx.exchange.is_valid_resolution(resolution):
+        valid_resolutions = ", ".join(ctx.exchange.resolutions.keys())
+        click.echo("Resolution {} is not supported.\n"
+                   "Please choose one of the following: {}".format(resolution,
+                                                                   valid_resolutions))
+        sys.exit(1)
+
+    # Initialise a strategy.
     strategy = STRATEGIES[strategy]()
-
-    if start:
-        start = datetime.datetime.strptime(start, "%Y-%m-%d %H:%M:%S")
-    if end:
-        end = datetime.datetime.strptime(end, "%Y-%m-%d %H:%M:%S")
-
-    if backtest:
-        if start is None or end is None:
-            click.echo("Error! For backtests you must provide a timeframe by setting start and end!")
-            sys.exit(1)
 
     bot = get_bot(market, strategy, resolution, start, end, btc, coins)
     bot.start(backtest, automatic)